+++ src/StardewModdingAPI/Program.cs
@@ -230,6 +230,7 @@ namespace StardewModdingAPI
         {
             // load settings
             this.Settings = JsonConvert.DeserializeObject<SConfig>(File.ReadAllText(Constants.ApiConfigPath));
+            this.GameInstance.VerboseLogging = this.Settings.VerboseLogging;
 
             // load core components
             this.ModRegistry = new ModRegistry(this.Settings.ModCompatibility);
@@ -266,6 +267,8 @@ namespace StardewModdingAPI
                 this.Monitor.Log($"You configured SMAPI to not check for updates. Running an old version of SMAPI is not recommended. You can enable update checks by reinstalling SMAPI or editing {Constants.ApiConfigPath}.", LogLevel.Warn);
             if (!this.Monitor.WriteToConsole)
                 this.Monitor.Log("Writing to the terminal is disabled because the --no-terminal argument was received. This usually means launching the terminal failed.", LogLevel.Warn);
+            if (this.Settings.VerboseLogging)
+                this.Monitor.Log("Verbose logging enabled.", LogLevel.Trace);
 
             // validate XNB integrity
             if (!this.ValidateContentIntegrity())